fix the issues: 1. replace tab with space in server_bin_log.lib.php 2. move $bin_log to server_binlog.php 3. add server common function file server_common.lib.php
