+++ server_binlog.php
@@ -12,12 +12,24 @@
 require_once 'libraries/common.inc.php';
 
 /**
- * Does the common work, provides $binary_logs
+ * Does the common work
  */
 require_once 'libraries/server_common.inc.php';
 
 require_once 'libraries/server_bin_log.lib.php';
 
+/**
+ * array binary log files
+ */
+$binary_logs = PMA_DRIZZLE
+    ? null
+    : $GLOBALS['dbi']->fetchResult(
+        'SHOW MASTER LOGS',
+        'Log_name',
+        null,
+        null,
+        PMA_DatabaseInterface::QUERY_STORE
+    );
 
 if (! isset($_REQUEST['log'])
     || ! array_key_exists($_REQUEST['log'], $binary_logs)
